+++ renderer/core-html-renderer/src/main/node/src/components/sidebar/Sidebar.svelte
@@ -2,6 +2,8 @@
     import ContextsItem from './components/Contexts-item.svelte';
     import TypeItem from './components/Type-item.svelte';
     import EnumItem from './components/Enum-item.svelte';
+    import EndpointIcon from './icons/endpoint.svg';
+    import ModelIcon from './icons/model.svg';
 
     const { apiContexts, usedApiTypes, usedEnumerations } = window.apiatorJson;
 
@@ -12,12 +14,20 @@
     };
 </script>
 
+<style>
+    .sidebar__group_icon {
+        width: 16px;
+    }
+</style>
+
 <div class="sidebar">
     <ul class="sidebar__group">
         <div
             on:click={handleGroupTitleClick}
             class="group__title js_sidebar-title-endpoints">
-            <i class="fa fa-plug" />
+            <i
+                style="background-image: url({EndpointIcon})"
+                class="sidebar__group_icon" />
             <span>Endpoints</span>
         </div>
         <li class="group__content group__content_active js_sidebar-endpoints">
@@ -31,7 +41,9 @@
         <div
             on:click={handleGroupTitleClick}
             class="group__title js_sidebar-title-model">
-            <i class="fa fa-puzzle-piece" />
+            <i
+                style="background-image: url({ModelIcon})"
+                class="sidebar__group_icon" />
             <span>Types</span>
         </div>
         <li class="group__content js_sidebar-model">