+++ gradle-plugin/apiator-gradle-plugin/src/main/groovy/com/ainrif/apiator/gradle/ApiatorGradlePlugin.groovy
@@ -18,7 +18,6 @@ package com.ainrif.apiator.gradle
 import com.ainrif.apiator.gradle.internal.ApiatorInternalRunner
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.artifacts.repositories.MavenArtifactRepository
 import org.gradle.api.plugins.JavaPlugin
 import org.gradle.api.plugins.JavaPluginConvention
 import org.gradle.api.tasks.SourceSet
@@ -87,11 +86,15 @@ class ApiatorGradlePlugin implements Plugin<Project> {
 
     private Object configureDependencyConstraints(Project prj) {
         def apiatorVersion = ApiatorGradlePlugin.package.implementationVersion
-        prj.dependencies.constraints.add('compile', "com.ainrif.apiator:api:${apiatorVersion}")
 
-        prj.repositories.maven(hint(MavenArtifactRepository) {
-            url = "http://dl.bintray.com/ainrif/maven"
-        })
+        if (!apiatorVersion) {
+            prj.logger.lifecycle("DEV mode detected. Local Maven Repo was added. Version changed to {}", prj.version)
+            prj.repositories.mavenLocal()
+            apiatorVersion = prj.version
+        }
+
+        prj.repositories.mavenCentral()
+        prj.dependencies.constraints.add('compile', "com.ainrif.apiator:api:${apiatorVersion}")
 
         prj.configurations.maybeCreate('apiatorCompile').with {
             def apiatorVersionConstraints = apiatorModules.collect {