+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v0.13.x - Performance optimisations
 
 - new : home site apiator.info and doc page docs.apiator.info
 - improvement: "Reflections" classpath scanner was replaced with parallel ClassGraph
+- improvement: HTML bundle size was reduced by replacing jQuery and Lodash dependencies with native js code
 - improvement: CI & CD pipelines were split
 
 v0.12.x