+++ CHANGELOG.md
@@ -1,9 +1,11 @@
-v0.13.x - Performance optimisations
+v0.13.x - Performance optimizations
 
 - new : home site apiator.info and doc page docs.apiator.info
 - improvement: "Reflections" classpath scanner was replaced with parallel ClassGraph
 - improvement: HTML bundle size was reduced by replacing jQuery and Lodash dependencies with native js code
 - improvement: CI & CD pipelines were split
+- improvement: release deploy to maven central
+- improvement: release plugin to gradle portal
 
 v0.12.x